Move the clean logic to TizenProject
