+++ lib/commands/clean.dart
@@ -4,13 +4,9 @@
 
 // @dart = 2.8
 
-import 'package:file/file.dart';
-import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/commands/clean.dart';
-import 'package:flutter_tools/src/globals.dart' as globals;
 import 'package:flutter_tools/src/project.dart';
 import 'package:flutter_tools/src/runner/flutter_command.dart';
-import 'package:path/path.dart';
 
 import '../tizen_project.dart';
 
@@ -20,50 +16,9 @@ class TizenCleanCommand extends CleanCommand {
   /// See: [CleanCommand.runCommand] in `clean.dart`
   @override
   Future<FlutterCommandResult> runCommand() async {
-    final FlutterProject flutterProject = FlutterProject.current();
-    _cleanTizenProject(TizenProject.fromFlutter(flutterProject));
-
+    final TizenProject tizenProject =
+        TizenProject.fromFlutter(FlutterProject.current());
+    tizenProject.clean();
     return super.runCommand();
-  }
-
-  void _cleanTizenProject(TizenProject project) {
-    if (!project.existsSync()) {
-      return;
-    }
-    _deleteFile(project.managedDirectory);
-
-    if (project.isDotnet) {
-      _deleteFile(project.editableDirectory.childDirectory('bin'));
-      _deleteFile(project.editableDirectory.childDirectory('obj'));
-      _deleteFile(project.editableDirectory.childFile('Runner.csproj.user'));
-    } else {
-      if (project.isMultiApp) {
-        _deleteFile(project.uiAppDirectory.childDirectory('Debug'));
-        _deleteFile(project.uiAppDirectory.childDirectory('Release'));
-        _deleteFile(project.serviceAppDirectory.childDirectory('Debug'));
-        _deleteFile(project.serviceAppDirectory.childDirectory('Release'));
-      } else {
-        _deleteFile(project.editableDirectory.childDirectory('Debug'));
-        _deleteFile(project.editableDirectory.childDirectory('Release'));
-      }
-    }
-  }
-
-  /// Source: [CleanCommand.deleteFile] in `clean.dart` (simplified)
-  void _deleteFile(FileSystemEntity file) {
-    if (!file.existsSync()) {
-      return;
-    }
-    final String path = relative(file.path);
-    final Status status = globals.logger.startProgress(
-      'Deleting $path...',
-    );
-    try {
-      file.deleteSync(recursive: true);
-    } on FileSystemException catch (error) {
-      globals.printError('Failed to remove $path: $error');
-    } finally {
-      status?.stop();
-    }
   }
 }
